move scheduler into tools
