+++ builder.js
@@ -32,6 +32,7 @@ module.exports = (env) => {
         './src/ts/admin.ts',
         './src/ts/profile.ts',
         './src/ts/edit.ts',
+        './src/ts/scheduler.ts',
         './src/ts/team.ts',
         './src/ts/metadata.ts',
         './src/ts/uploads.ts',
