+++ src/templates/head.html
@@ -82,6 +82,7 @@
               {{ 'Tools'|trans }}
               </div>
               <div class='dropdown-menu'>
+                <a class='dropdown-item' href='scheduler.php'><i class='fas fa-calendar-days fa-fw'></i>{{ 'Scheduler'|trans }}</a>
                 <a class='dropdown-item' href='compounds.php'><i class='fas fa-flask fa-fw'></i>{{ 'Compounds'|trans }}</a>
                 <a class='dropdown-item' href='chem-editor.php'><i class='fas fa-pen-to-square fa-fw'></i> {{ 'Chemical Structure Editor'|trans }}</a>
                 <a class='dropdown-item' href='syc.php'><i class='fas fa-dna fa-fw'></i> {{ 'DNA Cloning'|trans }}</a>
