+++ src/templates/ucp.html
@@ -115,14 +115,8 @@
   <h2 id='schedulerSettings' class='p-2 pl-3 section-title'>{{ 'Scheduler'|trans }}</h2>
 
   <div class='pl-3 mt-2 mb-5'>
-    <div class='d-flex justify-content-between'>
-      {# setting to toggle the display of week-ends on teams scheduler tab https://fullcalendar.io/docs/weekends #}
-      <label for='showWeekends' class='col-form-label'>{{ 'Show week ends'|trans }}</label>
-      <label class='switch ucp-align'>
-        <input id='showWeekends' type='checkbox' data-trigger='change' data-model='users/me' data-target='show_weekends' {{ App.Users.userData.show_weekends == 1 ? 'checked="checked"' }}>
-        <span class='slider'></span>
-      </label>
-    </div>
+    {# setting to toggle the display of week-ends on teams scheduler tab https://fullcalendar.io/docs/weekends #}
+    {% include 'binary-setting.html' with {'label': 'Show week ends'|trans, 'model': 'users/me', 'slug': 'show_weekends', 'src': App.Users.userData} %}
   </div>
 
   {# KEYBOARD SHORTCUTS #}
