+++ tests/cypress/integration/team.cy.ts
@@ -13,24 +13,23 @@ describe('Team', () => {
     cy.htmlvalidate();
 
     // Scheduler with selected item
-    cy.visit('/team.php?tab=1&item=1');
-    cy.get('[data-tabtarget="1"]').should('have.class', 'selected');
+    cy.visit('/scheduler.php?item=1');
     cy.get('[data-action="remove-param-reload"][data-target="item"]').should('exist');
+    cy.get('#loading-spinner').should('not.exist');
+    cy.htmlvalidate();
+
+    // Members
+    cy.visit('/team.php?tab=1');
+    cy.get('[data-tabtarget="1"]').should('have.class', 'selected');
+    cy.get('h3.section-title').should('contain.text', 'Members');
     cy.get('#loading-spinner').should('not.exist');
     cy.get('div[data-tabcontent="1"]').htmlvalidate();
 
-    // Members
+    // Send email
     cy.visit('/team.php?tab=2');
     cy.get('[data-tabtarget="2"]').should('have.class', 'selected');
-    cy.get('h3.section-title').should('contain.text', 'Members');
-    cy.get('#loading-spinner').should('not.exist');
-    cy.get('div[data-tabcontent="2"]').htmlvalidate();
-
-    // Send email
-    cy.visit('/team.php?tab=4');
-    cy.get('[data-tabtarget="4"]').should('have.class', 'selected');
     cy.get('h3.section-title').should('contain.text', 'Send an email to users');
     cy.get('#loading-spinner').should('not.exist');
-    cy.get('div[data-tabcontent="4"]').htmlvalidate();
+    cy.get('div[data-tabcontent="2"]').htmlvalidate();
   });
 });
